refactor: rename new UI component
