+++ src/frontend/src/lib/components/ui/SplitPane.svelte
