+++ src/frontend/src/routes/(app)/+layout.svelte
@@ -6,7 +6,7 @@
 	import Modals from '$lib/components/core/Modals.svelte';
 	import Header from '$lib/components/hero/Header.svelte';
 	import Hero from '$lib/components/hero/Hero.svelte';
-	import MainLayout from '$lib/components/ui/MainLayout.svelte';
+	import SplitPane from '$lib/components/ui/SplitPane.svelte';
 	import { authNotSignedIn, authSignedIn } from '$lib/derived/auth.derived';
 	import { pageToken } from '$lib/derived/page-token.derived';
 	import { token } from '$lib/stores/token.store';
@@ -32,7 +32,7 @@
 	<Header back={route === 'settings'} />
 
 	<AuthGuard>
-		<MainLayout>
+		<SplitPane>
 			<div class="" slot="menu">
 				<!-- TODO: add menu-->
 			</div>
@@ -48,7 +48,7 @@
 			<LoadersGuard>
 				<slot />
 			</LoadersGuard>
-		</MainLayout>
+		</SplitPane>
 
 		<Modals />
 	</AuthGuard>
